requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
blinker==1.4
Brlapi==0.6.6
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
crcmod==1.7
cryptography==2.1.4
defer==1.0.6
distro-info===0.14ubuntu0.2
Django==3.2.12
djangorestframework==3.13.1
firebase-admin==3.0.0
Flask==0.12.2
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
Guake==3.0.5
httplib2==0.14.0
idna==2.8
importlib-metadata==1.6.0
itsdangerous==0.24
Jinja2==2.10
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
louis==3.5.0
MarkupSafe==1.0
msgpack==0.6.2
mysql-connector-python==2.1.6
mysqlclient==1.4.6
netifaces==0.10.4
olefile==0.45.1
onboard==1.4.1
paho-mqtt==1.4.0
path==13.2.0
pbr==3.1.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.10.0
psutil==5.4.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycairo==1.16.2
PyChromecast==0.8.1
pycrypto==2.6.1
PyGObject==3.26.1
pyinotify==0.9.6
PyMySQL==0.9.3
pyOpenSSL==17.5.0
python-apt===1.6.5-ubuntu0.5-elementary17-ubuntu5.1.7.1
pytz==2021.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
rsa==4.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.4.2
ssh-import-id==5.7
typing_extensions==4.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
uritemplate==3.0.0
urllib3==1.25.6
vboxapi==1.0
Werkzeug==0.14.1
xkit==0.0.0
youtube_dl==2018.3.14
zeroconf==0.19.1
zipp==3.1.0
